Bump dependent tslint version to 5.1.0 (fixes #354)
  * 'no-cookies' now explicitly require type checking (extends
     Lint.Rules.TypedRule)
  * The following rules now optionally require type checking (extend
    from Lint.Rules.OptionallyTypedRule):
      * no-string-based-set-immediate
      * no-string-based-set-interval
      * no-string-based-set-timeout
  * 'no-single-line-block-comment' no longer supports nested single
    line comments
  * Removed 'no-unused-imports' in favor of the TSLint implementation
  * The default severity for TSLint is now error; fixing easy violations that
    were exposed by that change
  * Add extended metadata for new rules provided by TSLint
