+++ .vscode/settings.json
@@ -2,6 +2,7 @@
 {
     "editor.insertSpaces": true,
     "editor.tabSize": 4,
+    "tslint.nodePath": "./node_modules/tslint/bin/tslint",
     "typescript.tsdk": "./node_modules/typescript/lib",
     "files.trimTrailingWhitespace": true
 }
