+++ package.json
@@ -33,16 +33,17 @@
     "grunt-contrib-copy": "0.8.0",
     "grunt-contrib-watch": "0.6.1",
     "grunt-mocha-test": "0.12.7",
-    "grunt-ts": "5.4.0",
-    "grunt-tslint": "4.0.1",
+    "grunt-ts": "5.5.1",
+    "grunt-tslint": "5.0.1",
     "load-grunt-tasks": "3.2.0",
     "mocha": "2.2.5",
     "time-grunt": "1.2.1",
-    "tslint": "4.4.2",
+    "tslint": "5.1.0",
     "typescript": "2.1.4",
     "underscore": "1.8.3"
   },
   "peerDependencies": {
-    "tslint": ">=4.0.0"
+    "tslint": ">=5.1.0",
+    "tsutils": "1.6.0"
   }
 }
