+++ src/noStringBasedSetImmediateRule.ts
@@ -7,7 +7,7 @@ import {ExtendedMetadata} from './utils/ExtendedMetadata';
 /**
  * Implementation of the no-string-parameter-to-function-call rule.
  */
-export class Rule extends Lint.Rules.AbstractRule {
+export class Rule extends Lint.Rules.OptionallyTypedRule {
 
     public static metadata: ExtendedMetadata = {
         ruleName: 'no-string-based-set-immediate',
@@ -25,12 +25,12 @@ export class Rule extends Lint.Rules.AbstractRule {
     };
 
     public apply(sourceFile : ts.SourceFile): Lint.RuleFailure[] {
-        const documentRegistry = ts.createDocumentRegistry();
-        const languageServiceHost = Lint.createLanguageServiceHost('file.ts', sourceFile.getFullText());
-        const languageService = ts.createLanguageService(languageServiceHost, documentRegistry);
+        return this.applyWithProgram(sourceFile, undefined);
+    }
 
+    public applyWithProgram(sourceFile : ts.SourceFile, program : ts.Program): Lint.RuleFailure[] {
         const walker : Lint.RuleWalker = new NoStringParameterToFunctionCallWalker(
-            sourceFile , 'setImmediate', this.getOptions(), languageService
+            sourceFile , 'setImmediate', this.getOptions(), program
         );
         return this.applyWithWalker(walker);
     }
