+++ test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutFailingTestInput-constructor.ts
@@ -1,0 +1,6 @@
+
+class MyClass {
+    constructor(arg1) {
+        setTimeout(arg1, 5);
+    }
+}
