+++ test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutFailingTestInput-template-strings.ts
@@ -1,0 +1,2 @@
+var data = 'alert(1)';
+window.setTimeout(`${data}`, 200);
