+++ test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutTestInput-formerFalsePositive.ts
@@ -1,0 +1,5 @@
+
+function invoke(functionArg1 : () => void, functionArg2 = () => {}) {
+    setTimeout(functionArg1);
+    setTimeout(functionArg2);
+}
