+++ test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutTestInput-functionParamFunction.ts
@@ -1,0 +1,3 @@
+function foo(callback: Function) {
+    setTimeout(callback, 0);
+}
