+++ tsconfig.json
@@ -2,6 +2,11 @@
     "compilerOptions": {
         "declaration": true,
         "jsx": "react",
+        "lib": [
+            "es6",
+            "dom",
+            "scripthost"
+        ],
         "module": "commonjs",
         "outDir": "dist/src",
         "removeComments": true,
@@ -16,7 +21,6 @@
         "src/chaiVagueErrorsRule.ts",
         "src/exportNameRule.ts",
         "src/fixNoRequireImportsFormatter.ts",
-        "src/fixNoUnusedImportsFormatter.ts",
         "src/fixNoVarKeywordFormatter.ts",
         "src/functionNameRule.ts",
         "src/importNameRule.ts",
@@ -58,7 +62,6 @@
         "src/noRelativeImportsRule.ts",
         "src/noReservedKeywordsRule.ts",
         "src/noSingleLineBlockCommentRule.ts",
-        "src/noSparseArraysRule.ts",
         "src/noStatelessClassRule.ts",
         "src/noStringBasedSetImmediateRule.ts",
         "src/noStringBasedSetIntervalRule.ts",
@@ -72,7 +75,6 @@
         "src/noUnnecessaryOverrideRule.ts",
         "src/noUnnecessarySemicolonsRule.ts",
         "src/noUnsupportedBrowserCodeRule.ts",
-        "src/noUnusedImportsRule.ts",
         "src/noUselessFilesRule.ts",
         "src/noVarSelfRule.ts",
         "src/noWithStatementRule.ts",
@@ -105,7 +107,6 @@
         "src/tests/ChaiVagueErrorsRuleTests.ts",
         "src/tests/ExportNameRuleTests.ts",
         "src/tests/FixNoRequireImportsFormatterTests.ts",
-        "src/tests/FixNoUnusedImportsFormatterTests.ts",
         "src/tests/FixNoVarKeywordFormatterTests.ts",
         "src/tests/FunctionNameRuleTests.ts",
         "src/tests/GetImplicitRoleTest.ts",
@@ -147,7 +148,6 @@
         "src/tests/NoRelativeImportsRuleTests.ts",
         "src/tests/NoReservedKeywordsTests.ts",
         "src/tests/NoSingleLineBlockCommentRuleTests.ts",
-        "src/tests/NoSparseArraysRuleTests.ts",
         "src/tests/NoStatelessClassRuleTests.ts",
         "src/tests/NoStringBasedSetImmediateTests.ts",
         "src/tests/NoStringBasedSetIntervalTests.ts",
@@ -159,7 +159,6 @@
         "src/tests/NoUnnecessaryOverrideRuleTests.ts",
         "src/tests/NoUnnecessarySemicolonsTests.ts",
         "src/tests/NoUnsupportedBrowserCodeRuleTests.ts",
-        "src/tests/NoUnusedImportsTests.ts",
         "src/tests/NoUselessFilesRuleTests.ts",
         "src/tests/NoVarSelfRuleTests.ts",
         "src/tests/NoWithStatementTests.ts",
