+++ tsconfig.testdata.json
@@ -18,6 +18,7 @@
         "test-data/ExportName/ExportNameRulePassingTestInput2.tsx",
         "test-data/NoCookies/NoCookiesFailingTestInput.ts",
         "test-data/NoCookies/NoCookiesTestInput-error.ts",
+        "test-data/NoCookies/NoCookiesPassingTestInput.ts",
         "test-data/NoFunctionConstructorWithStringArgsTestInput.ts",
         "test-data/NoOctalLiteral/NoOctalLiteralTestInput-passing.ts",
         "test-data/NoReservedKeywords/NoReservedKeywordsTestInput-any.ts",
@@ -80,12 +81,23 @@
         "test-data/NoReservedKeywords/NoReservedKeywordsTestInput-while.ts",
         "test-data/NoReservedKeywords/NoReservedKeywordsTestInput-with.ts",
         "test-data/NoReservedKeywords/NoReservedKeywordsTestInput-yield.ts",
+        "test-data/NoStringBasedSetImmediateTestInput.ts",
+        "test-data/NoStringBasedSetIntervalTestInput.ts",
+        "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutFailingTestInput-constructor.ts",
+        "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutFailingTestInput-parameters.ts",
+        "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutFailingTestInput-template-strings.ts",
         "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutFailingTestInput-any-functions.ts",
         "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutFailingTestInput-any-variables.ts",
         "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutFailingTestInput-string-functions.ts",
         "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutFailingTestInput-string-literals.ts",
         "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutFailingTestInput-string-variables.ts",
         "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutFailingTestInput-walker-error.ts",
+        "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutTestInput-functionParamFunction.ts",
+        "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutTestInput-shadowedVariables.ts",
+        "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutTestInput-formerFalsePositive.ts",
+        "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutTestInput-issue46.ts",
+        "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutTestInput-case2.ts",
+        "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutTestInput-case3.ts",
         "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutTestInput-error.ts",
         "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutTestInput-error3.ts",
         "test-data/NoStringBasedSetTimeout/NoStringBasedSetTimeoutTestInput-error4.ts",
